Add a xla flag `--xla_force_execution_mode` to control execution mode on the CPU backend.
Fix a bug on CPU when we async dispatch a computation with python callback and arg donation by committing donation early: when `TfrtCpuBuffer` is calling its destructor, it will wait for pending donation, blocking the thread and holding the python gil, but right now donation will be committed after computation finishes which requires to hold the python gil as well because it has python callback, this leads to deadlocking. Demonstrated in `force_async_dispatch_on_cpu_test.py`.

PiperOrigin-RevId: 619700618